v2.8.2

+++ src/ltrace/ltrace/slicer/tests/ltrace_plugin_test.py
@@ -146,10 +146,16 @@ class LTracePluginTest(qt.QObject, ScriptedLoadableModule.ScriptedLoadableModule
             # Disconnect signals that keep the widget's method alive
             self._module_widget.cleanup()
 
+        # Close scene before attempt to delete the module widget instance
+        if self.__after_clear:
+            self.__close_project()
+
+        if self._module_widget is not None:
             # Delete the module widget immediately after processing events.
             # deleteLater() only deletes after the tests are finished.
             slicer.app.processEvents(1000)
-            self._module_widget.parent.delete()
+            if self._module_widget.parent != slicer.modules.AppContextInstance.mainWindow:
+                self._module_widget.parent.delete()
             slicer.app.processEvents(1000)
 
             weak_widget = weakref.ref(self._module_widget)
